Move data_keywords creation for data tables into document enhancers
This means we don't need to update every data table in push_data_keywords
which speeds up indexing.
